% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliary_Functions.R
\name{calc_parameter}
\alias{calc_parameter}
\title{calc_parameter}
\usage{
get_parameter(paramList, genotype)
}
\arguments{
\item{paramList}{sublist from referenceList, created by Make_reference functions.}

\item{genotype}{Genotype of interest}
}
\value{
double
}
\description{
Gets the genotype-dependent parameter of choice
}
\details{
This function gets whatever genotype-dependent factor desired. The inputs
are whatever parameter list of interest, then the genotype it is applied to.
}
\examples{
set.seed(42)
JaredDirichlet(n=4, alpha = c(0.1,0.2,0.3,0.4))

}
