% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliary_Functions.R
\name{CreateMosquitoes_Defined_Genotype}
\alias{CreateMosquitoes_Defined_Genotype}
\title{Create new Mosquito objects}
\usage{
CreateMosquitoes_Defined_Genotype(genMos, numMos, minAge, maxAge, ageDist)
}
\arguments{
\item{genMos}{List of genotypes for new Mosquitoes}

\item{numMos}{Integer number of Mosquitoes to create}

\item{minAge}{Integer specifying the minimum age of Mosquitoes}

\item{maxAge}{Integer specifying the maximum age of Mosquiotes}

\item{ageDist}{Distribution for ages of Mosquitoes. Must be length(maxAge-minAge+1)}
}
\value{
List of Mosquito objects length(sum(numMos))
}
\description{
Creates new Mosquito objects specifically for releases
}
\details{
This function creates new mosquitoes. It is similar to \code{\link{CreateMosquitoes_Eggs}}
and \code{\link{CreateMosquitoes_Distribution_Genotype}}, but intended for releases setup.
The assumption is that scientists can control the genotype of lab populations,
so this function takes a list of genotypes, then a vector of how many of each
genotype to release. There is some age variation, though that can be specified
precisely.
}
\examples{
set.seed(42)
genMos <- c("A","B","C")
numMos <- c(10,20,30)
minAge <- 1
maxAge <- 10
ageDist <- rep(x = 1, times = 10-1+1)/(10-1+1) #uniform distribution
CreateMosquitoes_Defined_Genotype(genMos, numMos, minAge, maxAge, ageDist)

set.seed(42)
genMos <- c("A","B","C")
numMos <- c(10,20,30)
minAge <- 5
maxAge <- 10
ageDist <- c(0,0,0,0,0,1) #all ages will be 10
CreateMosquitoes_Defined_Genotype(genMos, numMos, minAge, maxAge, ageDist)

}
