% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliary_Functions.R
\name{CreateMosquitoes_Eggs}
\alias{CreateMosquitoes_Eggs}
\title{Create new Mosquito objects}
\usage{
CreateMosquitoes_Eggs(genMos, numMos)
}
\arguments{
\item{genMos}{List of genotypes of Mosquitoes to create}

\item{numMos}{Vector specifying the number of each mosquito to create}
}
\value{
List of Mosquito objects length(sum(numMos))
}
\description{
Creates new Mosquito objects specifically for egg stage during reproduction
}
\details{
This function creates new mosquitoes. It is similar to \code{\link{CreateMosquitoes_Distribution_Genotype}}
and \code{\link{CreateMosquitoes_Defined_Genotype}}. It takes a list of genotypes
and the number of each to create. The difference is that this function always
creates them with age=0, as they are newly laid.
}
\examples{
CreateMosquitoes_Eggs(genMos = c("AA","BB"), numMos = c(10,20))

}
