% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_DaisyGeneratingFunction.R
\name{DaisyOffspring}
\alias{DaisyOffspring}
\title{Daisy Drive Offspring}
\usage{
DaisyOffspring(fGen, mGen, reference)
}
\arguments{
\item{fGen}{Female genotype}

\item{mGen}{Male genotype}

\item{reference}{Offspring reference list}
}
\value{
List(Alleles, Probabilities)
}
\description{
Create a list of offspring genotypes and their probabilities
}
\details{
Using the reference generated by \code{\link{MakeReference_DaisyDrive}},
this function expands the possible genotypes of the offspring and the ratios
that they occur. Similar to \code{\link{MultiplexOffspring_mLoci}} and
\code{\link{MultiplexOffspring_oLocus}}.
}
\examples{
ref <- MakeReferenceDaisy(H = 0.9, R = 0, S = 0, d = .001)
fGen <- "WW"
mGen <- "WW"

DaisyOffspring(fGen, mGen, ref)

}
