% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_MultiplexGeneratingFunction_multiLocus.R
\name{MakeReference_Multiplex_mLoci}
\alias{MakeReference_Multiplex_mLoci}
\title{mPlex Multiple Loci Offspring Reference}
\usage{
MakeReference_Multiplex_mLoci(H, R, S, d)
}
\arguments{
\item{H}{Vector of homing rates for each drive piece}

\item{R}{Vector of deleterious allele generation rates}

\item{S}{Vector of neutral allele generation rates}

\item{d}{Vector of background mutation rates at each locus}
}
\value{
list of homing and mendelian genotypes and rates
}
\description{
Create a list specifying the offspring probability distribution.
}
\details{
This function creates a reference list for \code{\link{MultiplexOffspring_mLoci}}.
Each drive element targets one locus. Each locus is independent of the rest.
The number of targets is specified by the length of H. R,S, and d must
be the same length as H, but will generally be the same number replicated that
many times. It is assumed that S is R/3, but this can be varied.
This function is similar to \code{\link{MakeReference_DaisyDrive}} and
\code{\link{MakeReference_Multiplex_oLocus}}
}
\examples{
H <- c(0.9,0.4,0.7) # This drive targets 3 loci
R <- c(0.001, 0.002, 0.003)
S <- c(0.0003, 0.0006, 0.001)
d <- c(0.0001, 0.0001, 0.0001)

MakeReference_Multiplex_mLoci(H,R,S,d)

}
