% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_MosquitoClass.R
\docType{class}
\name{Mosquito}
\alias{Mosquito}
\title{Mosquito Class Definition}
\format{An \code{\link{R6Class}} generator object}
\usage{
Mosquito
}
\description{
Mosquitos are the individual units being simulated in mPlex
}
\section{\strong{Constructor}}{

\itemize{
\item genotype: The genotype of the Mosquito to create
\item age: The age of the Mosquito at creation
}
}

\section{\strong{Methods}}{

\itemize{
\item set_age: Set the age of the Mosquito
\item set_mate: Set the mate of the Mosquito
\item set_genotype: Set the genotype of the Mosquito
\item get_age: Return the age of the Mosquito
\item get_mate: Return the mate of the Mosquito
\item get_genotype: Return the genotype of the Mosquito
\item age_one_day: Increase the age of the Mosquito by 1
\item print_male: Return age and genotype as comma separated values
\item print_female: Return age, genotype, and mate as comma separated values
}
}

\section{\strong{Fields}}{

\itemize{
\item age: Age of Mosquito
\item mate: Genotype of the mate of Mosquito
\item genotype: Genotype of Mosquito
}
}

\keyword{R6}
\keyword{class}
