% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_NetworkClass.R
\docType{class}
\name{Network}
\alias{Network}
\title{Network Class Definition}
\format{An \code{\link{R6Class}} generator object}
\usage{
Network
}
\description{
A Network is a collection of \code{\link{Patch}} with migration between them.
}
\section{\strong{Constructor}}{

\itemize{
\item networkParameters: see \code{\link{Network.Parameters}}
\item patchReleases: see \code{\link{Release_basicRepeatedReleases}}
\item reproductionType: one of c("DaisyDrive", "mPlex_oLocus", "mPlex_mLoci") specifying
the type of reproduction to use. Options are: \code{\link{DaisyOffspring}},
\code{\link{MultiplexOffspring_oLocus}}, or \code{\link{MultiplexOffspring_mLoci}}.
\item offspringReference: Inheritance patterns for the offspring. Must match reproductionType.
See \code{\link{MakeReference_DaisyDrive}}, \code{\link{MakeReference_Multiplex_oLocus}},
or \code{\link{MakeReference_Multiplex_mLoci}}.
\item migrationMale: a stochastic matrix whose dimensions conform to the number of patches
\item migrationFemale: a stochastic matrix whose dimensions conform to the number of patches
\item directory: character string specifying the output directory
}
}

\section{\strong{Methods}}{

\itemize{
\item get_nPatch: Return the number of patches in the Network
\item get_simTime: Return total time for the simulation
\item get_parallel: Return boolean if running parallel networks
\item get_moveVar: Return numeric variance in Dirchlet-Multinomial movement
\item get_runID: Return the run identifier
\item get_stageTime: Return duration of life stage
\item get_beta: Return wild-type daily fertility
\item get_mu: Return stage-specific density-independent death rate
\item get_alpha: Return larval density-dependent parameter
\item get_listPatch: Integer list of Patches in the Network
\item get_patch: Return a list of all patches or a single Patch specified by PatchID
\item get_reference: Return reference list for calculating offspring distribution
\item get_s: Return genotype-dependent list of fractional fertilities
\item get_tNow: Return current simulation time
\item get_directory: Return output directory
\item get_conADM: Return adult_male file connection
\item get_conADF: Return adult_female file connection
\item close_connections: Close conADM and conADF
\item get_migrationMale: Return the migrationMale matrix or a single row specified by PatchID
\item get_migrationFemale: Return the migrationFemale matrix or a single row specified by PatchID
\item get_migrationFractionMale: Return fraction of adult_male population that migrates, specified by PatchID
\item get_migrationFractionFemale: Return fraction of adult_female population that migrates, specified by PatchID
\item get_patchReleases: Return release schedule, specified by PatchID and sex
}
}

\section{\strong{Fields}}{

\itemize{
\item nPatch: Number of patches
\item simTime: Maximum time of simulation
\item parallel: Boolean if running several patches as the same time
\item moveVar: Variation in Dirichlet sampling for migration
\item runID: Run identifier
\item stageTime: Integer vector specifying the time spent in each life stage
\item beta: Integer wild-type daily fertility
\item mu: Integer vector specifying density-independent death rate for each life stage
\item alpha: Numeric vector specifying larval density factor for each patch
\item listPatch: Integer vector listing all the patches, 1:nPatch
\item patches: A list of \code{\link{Patch}} objects
\item reference: Inheritance pattern list
\item s: Fractional reduction/increase in genotype-dependent fertility
\item tNow: Current simulation time
\item directory: a character string of where to store output
\item conADM: a \code{\link[base]{connection}} to write male population dynamics out to
\item conAF1: a \code{\link[base]{connection}} to write female population dynamics out to
\item migrationMale: a stochastic matrix whose dimensions conform to the number of patches
\item migrationFractionMale: Numeric vector specifing the fraction of the population that migrates
\item migrationFemale: a stochastic matrix whose dimensions conform to the number of patches
\item migrationFractionFemale: Numeric vector specifing the fraction of the population that migrates
\item patchReleases: a list of release schedules for each patch
\item maleMoveOut: List of outgoing males per patch, exists for memory management
\item femaleMofeOut: List of outgoing females per patch, exists for memory management
\item maleMoveIn: List of incoming males per patch, exists for memory management
\item femaleMoveIn: List of incoming females per patch, exists for memory management
\item holdMale: Holder for use getting maleMoveIn. Prevents list issues with setting "NULL"
\item holdFemale: Holder for use getting femaleMoveIn. Prevents list issues with setting "NULL"
}
}

\keyword{R6}
\keyword{class}
