% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_PatchSimulation.R
\name{oneDay_Age_Patch}
\alias{oneDay_Age_Patch}
\title{Daily Aging}
\usage{
oneDay_Age_Patch(Population = NULL)
}
\arguments{
\item{Population}{The population to age}
}
\description{
Age all Mosquitoes in the population by 1 day
}
