% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_PatchSimulation.R
\name{oneDay_EggMature_Patch}
\alias{oneDay_EggMature_Patch}
\title{Daily Egg Maturation}
\usage{
oneDay_EggMature_Patch()
}
\description{
Eggs mature into larva. The average time spent as an egg is specified in
stageTime("E"). This is used to parametrize a lognormal function with
standard deviation log(1.2), slightly right-skewing the distribution.
}
