% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_PatchClass.R
\name{oneDay_larvaeReleases_Patch}
\alias{oneDay_larvaeReleases_Patch}
\title{Release larvae in a Patch}
\usage{
oneDay_larvaeReleases_Patch()
}
\description{
Based on this patch's release schedule, handle daily releases.
}
