% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_PatchSimulation.R
\name{oneDay_LarvalDeath_Patch}
\alias{oneDay_LarvalDeath_Patch}
\title{Daily Larvae Death}
\usage{
oneDay_LarvalDeath_Patch()
}
\description{
Kill larvae based on a chance of death specified by mu("L") and a density
dependend parameter alpha
}
