% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_PatchSimulation.R
\name{oneDay_PupaMature_Patch}
\alias{oneDay_PupaMature_Patch}
\title{Daily Pupa Maturation}
\usage{
oneDay_PupaMature_Patch()
}
\description{
Pupae mature into adults. The average time spent as a pupa is specified in
stageTime("P"). This is used to parametrize a lognormal function with
standard deviation log(1.2), slightly right-skewing the distribution. A binomial
distribtution with p=0.5 specifies the sex of maturing pupae. If male, pupa are
immediately added to the adult_male population. If female, pupa are added to
the unmated_female population.
}
