% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_PatchSimulation.R
\name{oneDay_Reproduction_Patch}
\alias{oneDay_Reproduction_Patch}
\title{Daily Egg Laying}
\usage{
oneDay_Reproduction_Patch()
}
\description{
The number of eggs laid per female is a Poisson distribution with a mean of
beta, the fertility. A multinomial is then used to distribute the number of
eggs laid over the offspring distribution (see \code{\link{DaisyOffspring}},
\code{\link{MultiplexOffspring_mLoci}}, or \code{\link{MultiplexOffspring_oLocus}}).
The new eggs are created by \code{\link{CreateMosquitoes_Eggs}}
}
