% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_NetworkClass.R
\name{oneRun_Network}
\alias{oneRun_Network}
\title{Run Simulation}
\usage{
oneRun_Network(conADM = NULL, conADF = NULL)
}
\arguments{
\item{conADM}{an optional \code{\link[base]{connection}} to write male population dynamics to}

\item{conADF}{an optional \code{\link[base]{connection}} to write female population dynamics to}
}
\description{
Run a single simulation on this network.
}
