% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliary_Functions.R
\name{Release_basicRepeatedReleases}
\alias{Release_basicRepeatedReleases}
\title{Make List of Mosquito Releases}
\usage{
Release_basicRepeatedReleases(releaseStart, releaseEnd, releaseInterval,
  releaseVector, sex = "M")
}
\arguments{
\item{releaseStart}{Day releases start}

\item{releaseEnd}{Day releases end}

\item{releaseInterval}{Interval between releases}

\item{releaseVector}{List of Mosquitoe objects to be releases}

\item{sex}{Character in c('M', 'F', 'L')}
}
\value{
List of release dates and the population to be released on that day.
}
\description{
Sets up a release schedule for a single patch, returns a list to be used in
\code{\link{oneDay_maleReleases_Patch}}, \code{\link{oneDay_femaleReleases_Patch}},
or \code{\link{oneDay_larvaeReleases_Patch}}.
}
\details{
See \code{\link{CreateMosquitoes_Defined_Genotype}} for how to setup
the release vector.
}
\examples{
# to setup for 3 patches but only release in the first with a defined release schedule:

patchReleases = replicate(n = 3,
                          expr = list(maleReleases = NULL,femaleReleases = NULL,larvaeReleases=NULL),
                          simplify = FALSE)

releaseMosquitoes <- CreateMosquitoes_Defined_Genotype(genMos, numMos, minAge, maxAge, ageDist)

patchReleases[[1]]$femaleReleases = Release_basicRepeatedReleases(releaseStart = 5,
releaseEnd = 30,
releaseInterval = 5,
releaseVector = releaseMosquitoes,
sex = "F")

patchReleases[[1]]$maleReleases = Release_basicRepeatedReleases(releaseStart = 50,
releaseEnd = 60,
releaseInterval = 1,
releaseVector = releaseMosquitoes,
sex = "M")

patchReleases[[1]]$larvaeReleases = Release_basicRepeatedReleases(releaseStart = 1,
releaseEnd = 5,
releaseInterval = 1,
releaseVector = releaseMosquitoes,
sex = "L")

}
